+++ src/SoftTissue.Core/NumericalMethods/Derivative/Derivative.cs
@@ -10,16 +10,6 @@ namespace SoftTissue.Core.NumericalMethods.Derivative
         public double Calculate<TInput>(Func<TInput, double, double> Equation, TInput input, double time)
             where TInput : ViscoelasticModelInput
         {
-            //double stepTime;
-            //if (variable == 0)
-            //{
-            //    stepTime = Math.Sqrt(Epsilon);
-            //}
-            //else
-            //{
-            //    stepTime = Math.Sqrt(Epsilon) * variable;
-            //}
-
             double actualValue = Equation(input, time);
             double nextValue = Equation(input, time + input.TimeStep);
 
@@ -28,20 +18,15 @@ namespace SoftTissue.Core.NumericalMethods.Derivative
 
         public double Calculate(Func<double, double> Equation, double step, double time)
         {
-            //double stepTime;
-            //if (variable == 0)
-            //{
-            //    stepTime = Math.Sqrt(Epsilon);
-            //}
-            //else
-            //{
-            //    stepTime = Math.Sqrt(Epsilon) * variable;
-            //}
-
             double actualValue = Equation(time);
             double nextValue = Equation(time + step);
 
             return (nextValue - actualValue) / step;
+        }
+
+        public double Calculate(double initialPoint, double finalPoint, double step)
+        {
+            return (finalPoint - initialPoint) / step;
         }
     }
 }