+++ src/SoftTissue.Core/Operations/Experimental/AnalyzeAndExtendResults/IAnalyzeAndExtendResults.cs
@@ -1,0 +1,10 @@
+﻿using SoftTissue.Core.Operations.Base;
+using SoftTissue.DataContract.Experimental.AnalyzeAndExtendResults;
+
+namespace SoftTissue.Core.Operations.Experimental
+{
+    /// <summary>
+    /// It is responsible to analyze and predict the experimental results.
+    /// </summary>
+    public interface IAnalyzeAndExtendResults : IOperationBase<AnalyzeAndExtendResultsRequest, AnalyzeAndExtendResultsResponse, AnalyzeAndExtendResultsResponseData> { }
+}