+++ src/SoftTissue.DataContract/Experimental/AnalyzeAndExtendResults/AnalyzeAndExtendResultsResponse.cs
@@ -1,0 +1,9 @@
+﻿using SoftTissue.DataContract.OperationBase;
+
+namespace SoftTissue.DataContract.Experimental.AnalyzeAndExtendResults
+{
+    /// <summary>
+    /// It represents the response content for AnalyzeAndPredictResults operations.
+    /// </summary>
+    public class AnalyzeAndExtendResultsResponse : OperationResponseBase<AnalyzeAndExtendResultsResponseData> { }
+}